Unifry message handling a little bit. This revolves around MSG_DATA
For the most part, MSG_DATA is fine as it is, but we need to be able
to create it from user data. That's implemented now.

As a result of this, though, it was worthwhile giving message_data the
same serialize() and extract_features() functions as other messages,
even if the implementation remains the same.

What changed, however, was that we encode payload size rather than
message buffer size in MSG_DATA. This is to do with the variable size
encoding of the size; in order to determine the encoding length of
this size, we have to first serialize it, and that may in turn impact
the encoding length. The payload size on the other hand is known at
the point of serialization.